timerqueue: Split out timerqueue_types.h
mainline inclusion
from mainline-v6.8-rc1
category: performace

Trimming down sched.h dependencies: timerqueue_types can include just
rbtree_types.h instead of pulling in rbtree.h.

Cc: Thomas Gleixner <tglx@linutronix.de>
Conflicts:
	include/linux/hrtimer_types.h
(cherry picked from commit eee51b0ae5c52a77ed65ad59b55002d1397b40d5)
